33.Medical_ChatBot_Project/Medical_ChatBot-main/requirements.txt
ctransformers
sentence-transformers
pinecone-client
langchain
flask
pypdf
python-dotenv
langchain-community
langchain-google-genai
protoc_gen_openapiv2
langchain_pinecone

-e .